backend/requirements.txt
Django==3.2.20
django_paypal==2.0
django_sslserver==0.22
dj-database-url==1.2.0
django-cors-headers==4.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
paypal==1.2.5
python-dotenv==0.21.0
Pillow==8.3.2
pytest==7.4.0
gunicorn==20.1.0
whitenoise==5.3.0
psycopg2-binary==2.9.5